backend/pyproject.toml
[project]
name = "audiobook-downloader-api"
version = "1.0.0"
description = "FastAPI backend for audiobook downloading"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "sse-starlette>=1.8.0",
    "aiosqlite>=0.19.0",
    "httpx>=0.26.0",
    # From existing tokybook requirements
    "requests>=2.32.0",
    "beautifulsoup4>=4.13.0",
    "lxml>=5.0.0",
    "mutagen>=1.47.0",
    "rich>=13.0.0",
    "yt-dlp>=2024.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "scrapers"]
